requirements.txt
absl-py @ file:///opt/concourse/worker/volumes/live/369745d3-6991-42a3-71cb-795aa3d79471/volume/absl-py_1615411186320/work
appnope @ file:///opt/concourse/worker/volumes/live/4f734db2-9ca8-4d8b-5b29-6ca15b4b4772/volume/appnope_1606859466979/work
astor==0.8.1
autopep8 @ file:///tmp/build/80754af9/autopep8_1620866417880/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
bcrypt==3.2.2
beautifulsoup4==4.6.0
cached-property @ file:///tmp/build/80754af9/cached-property_1600785575025/work
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_83242e7e-f82d-4a71-8ef2-9d71d212d249gu_wxmeq/croots/recipe/certifi_1655968827803/work/certifi
cffi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/croot-2p6q7e2r/cffi_1642701115063/work
chardet==3.0.4
click==8.0.1
cmake==3.21.0
colorama==0.4.4
coverage @ file:///opt/concourse/worker/volumes/live/4f1a5e30-f675-4496-6570-bc6669a34f2a/volume/coverage_1614614860186/work
cryptography @ file:///opt/concourse/worker/volumes/live/fab3474b-40c6-4001-48cb-802be615e43d/volume/cryptography_1633520388016/work
cycler==0.10.0
Cython @ file:///opt/concourse/worker/volumes/live/203a63af-5d69-4f3f-44a3-6968eb710a6a/volume/cython_1618435144870/work
DateTime==4.3
decorator @ file:///tmp/build/80754af9/decorator_1621259047763/work
Django==2.2.5
djangorestframework @ file:///home/conda/feedstock_root/build_artifacts/djangorestframework_1639677796037/work
dlib==19.22.0
et-xmlfile==1.1.0
figlet==0.0.1
Flask==2.0.1
gast==0.2.2
geographiclib==1.52
geopy==2.2.0
google-pasta==0.2.0
google-trans-new==1.1.9
googletrans==4.0.0rc1
grpcio @ file:///opt/concourse/worker/volumes/live/4f7ebd3c-7c5e-48f7-6ff5-de127acdc0d8/volume/grpcio_1614884166576/work
h11==0.9.0
h2==3.2.0
h5py @ file:///opt/concourse/worker/volumes/live/1e59f611-f2ec-4284-6b66-85a1e0d94198/volume/h5py_1622088428057/work
hpack==3.0.0
hstspreload==2021.12.1
html-sanitizer==1.9.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata @ file:///opt/concourse/worker/volumes/live/4e25a0be-45cc-4b73-4314-4604f00e30a4/volume/importlib-metadata_1617877365451/work
imutils==0.5.3
ipykernel @ file:///opt/concourse/worker/volumes/live/73e8766c-12c3-4f76-62a6-3dea9a7da5b7/volume/ipykernel_1596206701501/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///opt/concourse/worker/volumes/live/f33fa11b-d908-43e8-693a-f4b58d8c695c/volume/ipython_1617120878195/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
itsdangerous==2.0.1
jedi==0.17.0
Jinja2==3.0.1
joblib==1.0.1
JPype1==1.3.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1616770841739/work
jupyter-core @ file:///opt/concourse/worker/volumes/live/a699b83f-e941-4170-5136-bf87e3f37756/volume/jupyter_core_1612213304212/work
Keras-Applications @ file:///tmp/build/80754af9/keras-applications_1594366238411/work
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
kiwisolver @ file:///opt/concourse/worker/volumes/live/1e6d28e8-da14-4365-7344-567a05db7efc/volume/kiwisolver_1612282399491/work
konlpy==0.5.2
kss==2.5.1
lxml==4.6.3
Markdown @ file:///opt/concourse/worker/volumes/live/255f8945-2411-4469-7c82-a749af3aef33/volume/markdown_1614363822772/work
MarkupSafe==2.0.1
matplotlib @ file:///opt/concourse/worker/volumes/live/41e8cd50-031f-4dda-5787-dd3c4f4e0f08/volume/matplotlib-suite_1613407855571/work
mkl-fft==1.3.0
mkl-random @ file:///opt/concourse/worker/volumes/live/45391dab-1d51-4226-600a-9065613b835f/volume/mkl_random_1618853975662/work
mkl-service==2.3.0
mysqlclient @ file:///opt/concourse/worker/volumes/live/a00baea8-7042-4e61-5922-65c2a6fb6552/volume/mysqlclient_1609786256659/work
nltk==3.6.2
numpy @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_10c245a1-1c81-453c-8b70-3e944787a8d9mo1xhesu/croots/recipe/numpy_and_numpy_base_1653915529188/work
oauthlib==3.1.1
olefile==0.46
opencv-python==4.2.0.32
opencv-python-headless==4.5.3.56
openpyxl==3.0.7
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
pandas==1.2.4
parso @ file:///tmp/build/80754af9/parso_1617223946239/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow @ file:///opt/concourse/worker/volumes/live/0aff898a-aba9-4321-76c2-58d629e96356/volume/pillow_1617386155885/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1616415428029/work
protobuf==3.14.0
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pycodestyle @ file:///tmp/build/80754af9/pycodestyle_1615748559966/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyfiglet==0.8.post1
Pygments @ file:///tmp/build/80754af9/pygments_1621606182707/work
PyJWT @ file:///opt/concourse/worker/volumes/live/639caf1e-1cab-4e0c-78eb-5ad331abaa87/volume/pyjwt_1619651617491/work
PyMySQL @ file:///opt/concourse/worker/volumes/live/bb2ed9f0-4942-4298-76ba-b02a97cb1970/volume/pymysql_1610482900108/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
PySocks==1.7.1
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
pytz @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_1caf090c-01b6-4973-adac-d2eb884a674e7zonb9eb/croots/recipe/pytz_1654762641285/work
pyzmq==20.0.0
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rfc3986==1.5.0
schedule==0.6.0
scikit-learn==0.24.2
scipy @ file:///opt/concourse/worker/volumes/live/719943d5-6e16-41f8-68c4-f5820b33d300/volume/scipy_1641557782170/work
six @ file:///opt/concourse/worker/volumes/live/f983ba11-c9fe-4dff-7ce7-d89b95b09771/volume/six_1605205318156/work
sklearn==0.0
sniffio==1.2.0
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.0
termcolor==1.1.0
threadpoolctl==2.1.0
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tornado @ file:///opt/concourse/worker/volumes/live/d531d395-893c-4ca1-6a5f-717b318eb08c/volume/tornado_1606942307627/work
tqdm==4.61.1
traitlets @ file:///home/ktietz/src/ci/traitlets_1611929699868/work
tweepy==3.10.0
typing-extensions @ file:///home/ktietz/src/ci_mi/typing_extensions_1612808209620/work
urllib3==1.26.5
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
Werkzeug @ file:///opt/conda/conda-bld/werkzeug_1645628268370/work
wrapt==1.12.1
xlrd==2.0.1
zipp @ file:///tmp/build/80754af9/zipp_1615904174917/work
zope.interface==5.4.0
